Fixes: NB#268169 - FT_CN: The wrong behaviors happens when downloading MMS via manual
In some tricky scenario EventModels are not notifyed about event modfications because
TextChannelListener object may be destroyed while TrackerIO's transaction is not finished.
As result, it never recvs 'commited' signal and does not sends DBus notifications.

It happens because of TextChannelListener::m_EventTokens is not properly updated
when MMS notification is replaced with real message.

In some trickey scenarion Telepathy Channel is closed and TextChannelListener::finishedWithError() is called
hasPendingOperations() returns false and object is destroyed.