[commhistory-daemon] Track accumulated voice call counters.
Accumulated voice call counters is tracked separately from call history
to allow them to be individually reset or modified.

This implementation currently stores the accumulated voice call
counters in a settings file (~/.config/nemomobile/commhistoryd.conf).
They should probably be stored in the same database as the reset of
commhistory.

Provides a D-Bus interface for requesting the current value of the
counters and reseting the counters to 0.

The stored accumulated voice call counters only included completed
calls. The D-Bus service queries the commhistory to find the duration
of any currently ongoing calls and returns the total.