[commhistory-daemon] Remove OldDataDeleter
Contrary to the name, this actually only deletes old (90 days) calls.
The implementation is strange, the time is much too short, and the whole
concept is probably unnecessary when using a sane database. If it is
necessary to trim old data out of the database, it should be done more
generically with support for conversations as well, which is much more
important size/performance-wise anyway.