[commhistory-daemon] Remove error state for messages sent to offline contacts
For most supported services, offline messaging is possible and the
message will be sent when the remote user connects. It's inappropriate
to mark those messages as failed.

For services that don't do offline messaging, they should return an
error, which will separately set the status to failed.

A banner is displayed the first time you send a message to an offline
contact, reminding you that they are offline. This is not very clear and
should be improved as well.